requirements.txt
alembic==1.3.1
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.24.2
Werkzeug==0.16.0
WTForms==2.2.1
